Code Updating
Not a big change. OperationMC.java has been re-designed, but this design is still not finished, though it will not influence the whole system now. This attempt is for introducing an interactive environment. I have also introduced a new channel, namely ExpChannel6.java, which is a "real" sensorimotor channel placed in a 3-D simulation environment. This channel stands for the camera of a robot, which is placed in a limited world with several objects. You may find you are not able to run shell.jave anymore, since this new channel is loaded and you need to open a project in Webots, and of course, with a TON of configurations. I will upload the Webots I used and an instruction in the near future.