Code Updating
This is not a significant update, and I am also not satisfied with it to be honest. But it still provides something.
You can find a large range of updating since I have finally touched something fundamental, and it lets some hidden problems revealed.
First, for the updating, now it not only supports gathering information from the environment, but it also supports doing something to change the world. I planned to make this "changing" consciously, but this may need the current implement supports "reasoning on goals", which is not implemented yet. So, you may just find I manually gave NARS some orders to do. But please keep in mind that this is NOT what I want.
Secondly, for the hidden problems. Except for the deficiency on reasoning of goals, I also found the keyword "Operator" or "Operation" have different implementations and largely different functionalities. So, my next step is to clean these old and overlapping definitions. I have no other choices to continue.